refactor: refactor add fav logical
